requirements.txt
signalbot>=0.7.0
httpx>=0.27
aiosqlite>=0.20
python-dotenv>=1.0
fastapi>=0.115
uvicorn[standard]>=0.34
jinja2>=3.1
itsdangerous>=2.2
pytest>=8.0
